Update Decompilations Using ghidra_11.4.2

+++ ghidrecomps/bins/single-board-test-94a88b3ddedb802228f032ec1149b0bd/decomps/X509_NAME_hash-000c02a8.c
@@ -3,7 +3,7 @@ ulong X509_NAME_hash(X509_NAME *x)
 
 {
   EVP_MD *type;
-  ulong uVar1;
+  uint uVar1;
   size_t count;
   uchar *data;
   byte local_28;
