+++ ghidrecomps/bins/single-board-test-94a88b3ddedb802228f032ec1149b0bd/decomps/X509_subject_name_hash-000c01b8.c
@@ -3,7 +3,7 @@ ulong X509_subject_name_hash(X509 *x)
 
 {
   EVP_MD *type;
-  ulong uVar1;
+  uint uVar1;
   X509_NAME *a;
   size_t count;
   uchar *data;
