+++ ghidrecomps/bins/single-board-test-default-dd3d1850ff6a7f4acefac8c49c239c1d/decomps/main-00008b68.c
@@ -3,7 +3,7 @@ int main(void)
 
 {
   char local_50 [64];
-  ssize_t local_10;
+  int local_10;
   undefined4 local_c;
   
   local_c = 0xffffffff;
