+++ ghidrecomps/bins/single-board-test-default-dd3d1850ff6a7f4acefac8c49c239c1d/decomps/sha2_update-00016c08.c
@@ -3,7 +3,7 @@ void sha2_update(uint *param_1,void *param_2,uint param_3)
 
 {
   size_t __n;
-  uint local_1c;
+  size_t local_1c;
   void *local_18;
   uint local_c;
